+++ doc/source/changelog.md
@@ -25,6 +25,7 @@ Primarily developped with Godot 4.3.
 - Fixes
     - `VoxelBlockyModelMesh`: Fixed materials present directly in the mesh resource were not applied (only overrides in the model or on the terrain were applied)
     - `VoxelBlockyType`: Fixed configuration warning about missing variants when there is a base model specified
+    - `VoxelGeneratorGraph`: Fixed crash when using the `Image` node with a non-square image
     - `VoxelStreamSQLite`: 
         - Fixed `set_key_cache_enabled(true)` caused nothing to load
         - Fixed slow loading when the database path contains `res://` or `user://`